Refactor AccountFixture
- We deduplicated the fixture across the related modules and their tests
- It works with UUID instead of String converted to ByteArray
- Documentation added
