+++ darkside-test-lib/build.gradle.kts
@@ -25,8 +25,8 @@ dependencies {
     implementation(libs.androidx.multidex)
     implementation(libs.bundles.grpc)
 
+    androidTestImplementation(projects.sdkIncubatorLib)
     androidTestImplementation(libs.bundles.androidx.test)
-
     androidTestImplementation(libs.zcashwalletplgn)
     androidTestImplementation(libs.bip39)
 }
